// src/types.ts
export type TabType = 
  | 'beranda'
  | 'semua-turnamen'
  | 'ff'
  | 'mlbb'
  | 'total-hadiah'
  | 'pengumuman'
  | 'riwayat'
  | 'status-pembayaran'
  | 'form-pendaftaran'
  | 'pengaturan-umum'
  | 'bantuan'
  | 'arsip'
  | 'bagikan'
  | 'cara-daftar'
  | 'aturan'
  | 'tim'
  | 'info-match'
  | 'grup'
  | 'laporan'
  | 'kontak'
  | 'login'
  | 'prediksi'
  | 'saldo'
  | 'admin'
  | 'topup-game'
  | 'unduh-apk'
  | 'donasi'
  | 'gemini-ai';